requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
langchain==0.0.352
langchain-community==0.0.20
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
openai==1.3.7
tiktoken==0.5.1 
transformers